+++ src/payment/mod.rs
@@ -17,11 +17,13 @@ pub(crate) mod store;
 mod unified;
 
 pub use bolt11::Bolt11Payment;
+pub(crate) use bolt11::PaymentMetadata;
 pub use bolt12::Bolt12Payment;
 pub use onchain::OnchainPayment;
 pub use pending_payment_store::PendingPaymentDetails;
 pub use spontaneous::SpontaneousPayment;
 pub use store::{
-	ConfirmationStatus, LSPFeeLimits, PaymentDetails, PaymentDirection, PaymentKind, PaymentStatus,
+	ConfirmationStatus, LSPS2Parameters, PaymentDetails, PaymentDirection, PaymentKind,
+	PaymentStatus,
 };
 pub use unified::{UnifiedPayment, UnifiedPaymentResult};